chore(blockchains): initial version, needs 100% coverage
