+++ packages/blockchains/jest.setup.js
@@ -1,1 +1,5 @@
-jest.setTimeout(80000)
+jest.mock('@react-native-async-storage/async-storage', () =>
+  require('@react-native-async-storage/async-storage/jest/async-storage-mock'),
+)
+
+jest.setTimeout(60000)
