+++ packages/blockchains/package.json
@@ -136,6 +136,13 @@
       "<rootDir>/jest.setup.js"
     ]
   },
+  "dependencies": {
+    "@yoroi/api": "1.0.0",
+    "@yoroi/common": "1.5.4",
+    "@yoroi/explorers": "1.0.2",
+    "@yoroi/portfolio": "1.0.3",
+    "@yoroi/types": "1.5.10"
+  },
   "devDependencies": {
     "@commitlint/config-conventional": "^17.0.2",
     "@react-native-async-storage/async-storage": "^1.19.3",
@@ -147,9 +154,6 @@
     "@types/jest": "^29.5.12",
     "@types/react": "^18.2.55",
     "@types/react-test-renderer": "^18.0.7",
-    "@yoroi/common": "^1.5.4",
-    "@yoroi/portfolio": "1.0.3",
-    "@yoroi/types": "1.5.10",
     "bignumber.js": "^9.0.1",
     "commitlint": "^17.0.2",
     "del-cli": "^5.0.0",
@@ -173,9 +177,6 @@
   },
   "peerDependencies": {
     "immer": "^10.0.2"
-  },
-  "optionalDependencies": {
-    "react": ">= 16.8.0 <= 19.0.0"
   },
   "packageManager": "yarn@1.22.21",
   "engines": {
