+++ packages/blockchains/src/index.ts
@@ -1,7 +1,6 @@
 export * from './network-manager'
 export * from './types'
 
-export * from './helpers/date-to-epoch-info'
-export * from './helpers/epoch-progress'
-export * from './helpers/get-wallet-factory'
-export * from './helpers/to-chain-supported-network'
+export * from './cardano/helpers/date-to-epoch-info'
+export * from './cardano/helpers/epoch-progress'
+export * from './cardano/helpers/to-chain-supported-network'
