+++ packages/blockchains/src/types.ts
@@ -1,5 +1,5 @@
 import {Chain, Portfolio} from '@yoroi/types'
 
-export type NetworkTokenManagers = Readonly<
+export type TokenManagerByNetwork = Readonly<
   Record<Chain.SupportedNetworks, Portfolio.Manager.Token>
 >
